requirements/adm.backtoshops.com.requirements.txt
Django==1.5
Pillow==2.1.0
South==0.8.2
django-extensions==1.2.0
django-form-utils==1.0
django-formwizard==1.0
django-rosetta==0.7.4
django-misc==0.1.0
sorl-thumbnail==11.12
python-memcached==1.53
psycopg2==2.5.1
redis==2.8.0
M2Crypto==0.21.1.B2S
B2SCrypto==0.0.5
B2SProtocol==0.0.18
enum==0.4.4
ujson==1.33
B2SUtils==0.1.5
xmltodict==0.7.0
incf.countryutils==1.0
django-aloha-editor==1.0
pytz==2014.4
